backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
cryptography==41.0.7
PyJWT==2.10.1
anthropic==0.25.1
google-generativeai==0.8.6
pytesseract==0.3.10
pillow==9.5.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.23.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
